+++ internal/provider/resource_meraki_appliance_dns_split_profile_assignments_test.go
@@ -1,0 +1,132 @@
+// Copyright © 2024 Cisco Systems, Inc. and its affiliates.
+// All rights reserved.
+//
+// Licensed under the Mozilla Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://mozilla.org/MPL/2.0/
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: MPL-2.0
+
+package provider
+
+// Section below is generated&owned by "gen/generator.go". //template:begin imports
+import (
+	"fmt"
+	"os"
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/terraform"
+)
+
+// End of section. //template:end imports
+
+// Section below is generated&owned by "gen/generator.go". //template:begin testAcc
+
+func TestAccMerakiApplianceDNSSplitProfileAssignments(t *testing.T) {
+	if os.Getenv("APPLIANCE_DNS_SPLIT_PROFILE_ASSIGNMENTS") == "" {
+		t.Skip("skipping test, set environment variable APPLIANCE_DNS_SPLIT_PROFILE_ASSIGNMENTS")
+	}
+	if os.Getenv("TF_VAR_test_org") == "" || os.Getenv("TF_VAR_test_network") == "" {
+		t.Skip("skipping test, set environment variable TF_VAR_test_org and TF_VAR_test_network")
+	}
+	var checks []resource.TestCheckFunc
+
+	var steps []resource.TestStep
+	if os.Getenv("SKIP_MINIMUM_TEST") == "" {
+		steps = append(steps, resource.TestStep{
+			Config: testAccMerakiApplianceDNSSplitProfileAssignmentsPrerequisitesConfig + testAccMerakiApplianceDNSSplitProfileAssignmentsConfig_minimum(),
+		})
+	}
+	steps = append(steps, resource.TestStep{
+		Config: testAccMerakiApplianceDNSSplitProfileAssignmentsPrerequisitesConfig + testAccMerakiApplianceDNSSplitProfileAssignmentsConfig_all(),
+		Check:  resource.ComposeTestCheckFunc(checks...),
+	})
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheck(t) },
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps:                    steps,
+	})
+}
+
+// End of section. //template:end testAcc
+
+// Section below is generated&owned by "gen/generator.go". //template:begin importStateIdFunc
+
+func merakiApplianceDNSSplitProfileAssignmentsImportStateIdFunc(resourceName string) resource.ImportStateIdFunc {
+	return func(s *terraform.State) (string, error) {
+		primary := s.RootModule().Resources[resourceName].Primary
+		OrganizationId := primary.Attributes["organization_id"]
+
+		return fmt.Sprintf("%s", OrganizationId), nil
+	}
+}
+
+// End of section. //template:end importStateIdFunc
+
+// Section below is generated&owned by "gen/generator.go". //template:begin testPrerequisites
+
+const testAccMerakiApplianceDNSSplitProfileAssignmentsPrerequisitesConfig = `
+variable "test_org" {}
+variable "test_network" {}
+data "meraki_organization" "test" {
+  name = var.test_org
+}
+resource "meraki_network" "test" {
+  organization_id = data.meraki_organization.test.id
+  name            = var.test_network
+  product_types   = ["appliance"]
+}
+resource "meraki_appliance_dns_split_profile" "test" {
+  organization_id = data.meraki_organization.test.id
+  name = "Default profile"
+  nameservers_addresses = ["12.1.10.1"]
+  hostnames = ["*.test1.com"]
+}
+
+`
+
+// End of section. //template:end testPrerequisites
+
+// Section below is generated&owned by "gen/generator.go". //template:begin testAccConfigMinimal
+
+func testAccMerakiApplianceDNSSplitProfileAssignmentsConfig_minimum() string {
+	config := `resource "meraki_appliance_dns_split_profile_assignments" "test" {` + "\n"
+	config += `  organization_id = data.meraki_organization.test.id` + "\n"
+	config += `  items = [{` + "\n"
+	config += `    network_id = meraki_network.test.id` + "\n"
+	config += `    profile_id = meraki_appliance_dns_split_profile.test.id` + "\n"
+	config += `  }]` + "\n"
+	config += `}` + "\n"
+	return config
+}
+
+// End of section. //template:end testAccConfigMinimal
+
+// Section below is generated&owned by "gen/generator.go". //template:begin testAccConfigAll
+
+func testAccMerakiApplianceDNSSplitProfileAssignmentsConfig_all() string {
+	config := `resource "meraki_appliance_dns_split_profile_assignments" "test" {` + "\n"
+	config += `  organization_id = data.meraki_organization.test.id` + "\n"
+	config += `  items = [{` + "\n"
+	config += `    network_id = meraki_network.test.id` + "\n"
+	config += `    profile_id = meraki_appliance_dns_split_profile.test.id` + "\n"
+	config += `  }]` + "\n"
+	config += `}` + "\n"
+	return config
+}
+
+// End of section. //template:end testAccConfigAll
+
+// Section below is generated&owned by "gen/generator.go". //template:begin testAccConfigAdditional
+
+// End of section. //template:end testAccConfigAdditional